7nov commit

+++ admin/includes/view_all_posts.php
@@ -18,6 +18,25 @@ if (isset($_POST['checkBoxArray'])) {
                 $update_delete_post = mysqli_query($connection, $query);
                 confirmQuery($update_delete_post);
                 break;
+            case 'clone':
+                $query = "SELECT * FROM posts WHERE post_id= '{$postValueId}' ";
+                $select_post_query = mysqli_query($connection, $query);
+
+                while ($row = mysqli_fetch_array($select_post_query)) {
+                    $post_title = $row['post_title'];
+                    $post_category_id = $row['post_category_id'];
+                    $post_date = $row['post_date'];
+                    $post_author = $row['post_author'];
+                    $post_status = $row['post_status'];
+                    $post_image = $row['post_image'];
+                    $post_tag = $row['post_tag'];
+                    $post_content = mysqli_real_escape_string($connection, $row['post_content']);
+                }
+                $query = "INSERT INTO posts(post_category_id, post_title, post_author, post_date, post_image, post_content, post_tag, post_status )";
+                $query .= "VALUES({$post_category_id}, '{$post_title}', '{$post_author}', now(), '{$post_image}', '{$post_content}', '{$post_tag}', '{$post_status}' ) ";
+                $copy_query = mysqli_query($connection, $query);
+                confirmQuery($copy_query);
+                break;
         }
     }
 }
@@ -31,6 +50,7 @@ if (isset($_POST['checkBoxArray'])) {
                     <option value="published">Publish</option>
                     <option value="draft">Draft</option>
                     <option value="delete">Delete</option>
+                    <option value="clone">Clone</option>
                 </select>
             </div>
             <div class="col-xs-4">
@@ -49,8 +69,10 @@ if (isset($_POST['checkBoxArray'])) {
                     <th>Tags</th>
                     <th>Comments</th>
                     <th>Date</th>
+                    <th>Posts</th>
                     <th>Edit</th>
                     <th>Delete</th>
+                    <th>Views</th>
 
                 </tr>
             </thead>
@@ -71,6 +93,7 @@ if (isset($_POST['checkBoxArray'])) {
                     $post_tag = $row['post_tag'];
                     $post_comment_count = $row['post_comment_count'];
                     $post_date = $row['post_date'];
+                    $post_view_count = $row['post_view_count'];
 
                     echo "<tr>";
                     ?>
@@ -93,12 +116,19 @@ if (isset($_POST['checkBoxArray'])) {
                         echo "<td>{$post_status}</td>";
                         echo "<td><img width='100' src='../images/$post_image' alt='image'></td>";
                         echo "<td>{$post_tag}</td>";
-                        echo "<td>{$post_comment_count}</td>";
+
+                        $query = "SELECT * FROM comments WHERE comment_post_id = $post_id";
+                        $send_comment_query = mysqli_query($connection, $query);
+                        $row = mysqli_fetch_array($send_comment_query);
+                        $comment_id = $row['comment_id'];
+                        $count_comments = mysqli_num_rows($send_comment_query);
+
+                        echo "<td><a href='post_comments.php?id=$post_id'>{$count_comments}</a></td>";
                         echo "<td>{$post_date}</td>";
                         echo "<td> <a href='../post.php?p_id={$post_id}'>View Post</a></td>";
                         echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-                        echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
-
+                        echo "<td><a onClick=\"javascript: return confirm('Are you sure want to delete this post?'); \" href='posts.php?delete={$post_id}'>Delete</a></td>";
+                        echo "<td><a href='posts.php?reset={$post_id}'>{$post_view_count}</a></td>";
                         echo "</tr>";
                     }
                     ?>
@@ -111,6 +141,15 @@ if (isset($_POST['checkBoxArray'])) {
         $get_post_id = $_GET['delete'];
         $query = "DELETE FROM posts WHERE post_id = {$get_post_id}";
         $delete_query = mysqli_query($connection, $query);
+        header("Location: posts.php");
+    }
+    ?>
+
+    <?php
+    if (isset($_GET['reset'])) {
+        $get_post_id = $_GET['reset'];
+        $query = "UPDATE posts SET post_view_count = 0 WHERE post_id =" . mysqli_real_escape_string($connection, $_GET['reset']) . " ";
+        $reset_query = mysqli_query($connection, $query);
         header("Location: posts.php");
     }
     ?>
